requirements.txt
# Core Flask
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
mysql-connector-python==8.1.0
PyMySQL==1.1.0
mysqlclient==2.2.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
Flask-MySQLdb==2.0.0

# Authentication & Security
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.0.1
email-validator==2.0.0
Werkzeug==2.3.7
itsdangerous==2.1.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
MarkupSafe==2.1.3
zipp==3.16.2

# Development
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
